Refactor Market and ProductDetails components and screens
