+++ app/src/main/java/com/delighted2wins/souqelkhorda/features/market/presentation/component/Market/SearchBar.kt
@@ -1,9 +1,11 @@
-package com.delighted2wins.souqelkhorda.features.market.presentation.component
+package com.delighted2wins.souqelkhorda.features.market.presentation.component.Market
 
+import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Search
+import androidx.compose.material3.AlertDialogDefaults.shape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
@@ -11,14 +13,9 @@ import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalLayoutDirection
-import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.OffsetMapping
-import androidx.compose.ui.text.input.TransformedText
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import com.delighted2wins.souqelkhorda.app.theme.Til
 
@@ -47,7 +44,9 @@ fun SearchBar(
         } else null,
         singleLine = true,
         shape = RoundedCornerShape(24.dp),
-        modifier = modifier.height(56.dp),
+        modifier = modifier
+            .height(56.dp)
+            .border(width = 1.dp, color = Til, shape = shape),
         colors = TextFieldDefaults.colors(
             focusedContainerColor = Color.White,
             unfocusedContainerColor = Color.White,
@@ -69,7 +68,6 @@ fun SearchBar(
 fun SearchBarPreviewLTR() {
     SearchBar(query = "", onQueryChange = {}, isRtl = false)
 }
-
 
 @Preview(showBackground = true)
 @Composable
